pin_c: split pin_location.cpp, debug print
